requirements.txt
django 
langchain 
openai-whisper 
openai 
scikit-learn
python-dotenv
langchain-community 
langchain-openai
pydub
-e .
